Move list parser functions
